Answers now in its own component
Count no longer works. Can only show all the answers to all the questions. If I try to introduce the count to answerList it all just breaks down.
